refactor: split api / websocket routes
