+++ Cargo.toml
@@ -14,6 +14,7 @@ anyhow = "1.0.98"
 tokio = { version = "1.46.1", features = ["full"] }
 http-body-util = "0.1.3"
 tower = "0.5.2"
+tower-sessions = "0.14.0"
 http = "1.3.1"
 time = { version = "0.3.41", features = ["serde"] }
 serde = { version = "1.0.219", features = ["derive"] }
@@ -25,3 +26,6 @@ url = "2.5.4"
 futures-util = "0.3.31"
 axum-test = { version = "17.3.0", features = ["pretty-assertions", "ws"] }
 log = "0.4.27"
+axum-oidc = "0.6.0"
+cookie = "0.18.1"
+env_logger = "0.11.8"
